Add files via upload
These are functions to extract cell-chain corrected data and build a dataframe. Also convert dates from Julian to Gregorian, and filenames to datetime indices. This code is rough, and may need tweaking to suit your needs. I would like to move the class list outside of the actual builder function so it can be passed as an input.